restructure documentation routing
subpage serves view component based on url parameters
