+++ src/routes/docs/index.svelte
@@ -1,0 +1,5 @@
+<script lang="ts">
+    import Welcome from "../../views/welcome.svelte";
+</script>
+
+<Welcome />
